requirements.txt
streamlit==1.36.0
python-dotenv==1.0.0
whisper==1.1.10
moviepy==1.0.3
elevenlabs==1.2.0
pydub==0.25.1
deep-translator==1.11.4